fixed minor issue in auto swerve initialization
