+++ src/main/java/com/team1816/season/auto/AutoModeManager.java
@@ -6,7 +6,6 @@ import com.team1816.lib.auto.modes.DoNothingMode;
 import com.team1816.season.auto.modes.DriveStraightMode;
 import com.team1816.season.auto.modes.LivingRoomMode;
 import com.team1816.season.auto.modes.TuneDrivetrainMode;
-import com.team1816.season.auto.paths.DriveToBalance_Balance;
 import edu.wpi.first.wpilibj.RobotState;
 import com.team1816.season.auto.modes.*;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
@@ -165,7 +164,7 @@ public class AutoModeManager {
         LIVING_ROOM,
         DRIVE_STRAIGHT,
         // 2023
-        PLACECONEAUTOBALANCE_BALANCE
+        PLACE_CONE_AUTO_BALANCE
     }
 
     /**
@@ -183,8 +182,8 @@ public class AutoModeManager {
                 return new TuneDrivetrainMode();
             case LIVING_ROOM:
                 return (new LivingRoomMode(color));
-            case PLACECONEAUTOBALANCE_BALANCE:
-                return (new PlaceConeAutoBalance_BalanceMode(color));
+            case PLACE_CONE_AUTO_BALANCE:
+                return (new PlaceConeAutoBalanceMode(color));
             default:
                 System.out.println("Defaulting to drive straight mode");
                 return new DriveStraightMode();
